+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### v0.4.0-alpha
 *  [\#3 [ExpressionLanguage] lessThan helper](https://github.com/rest-control/rest-control/issues/3)
+* "Each" expression language helper
 #### v0.3.0-alpha
 *  Posibility to validate list of items in response filters
 *  Implementation of jsonPath in jsonPath helper